+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Unreleased
   * Document deploy workflow
   * Add stack utility tests
   * Expand design for refactoring deploy command
+  * Refactor deploy command using helper functions
+  * Add tests for deploy helper functions
 
 1.10.2 / 2025-06-04
 ===================